CONFIGURE
I've designed a rudimentary configure script to verify the existence of required utilities, the build system and wxcast.  The script also attempts to set all of the required configure variables and write the appropriate files.

There's an attempt, if the build system and wxcast are NOT installed, to use 'sudo [yum|apt] install ..... ' and 'sudo pip3 install -r requirements.txt' to install them.  YMMV.

Testing of the configuration script in a secondary shell environment is currently being done.